# api/openapi.yaml
openapi: 3.0.3
info:
  title: WORLDC2 C2 API
  description: Command & Control REST API for red team operations
  version: 2.0.0
  contact:
    name: ruby570bocadito
  license:
    name: MIT
servers:
  - url: http://localhost:9090
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Status:
      type: object
      description: Authenticated operational telemetry (requires sessions:list)
      properties:
        status:
          type: string
          example: ok
        active_sessions:
          type: integer
          example: 3
        listeners:
          type: integer
          example: 4
        uptime_seconds:
          type: integer
          format: int64
          description: Seconds since server start
    Session:
      type: object
      properties:
        ID:
          type: string
          example: "a1b2c3d4e5f6..."
        AgentID:
          type: string
          example: "agent-12345678"
        Hostname:
          type: string
          example: "DESKTOP-ABC123"
        OS:
          type: string
          example: "windows"
        Arch:
          type: string
          example: "amd64"
        Username:
          type: string
          example: "user"
        IsAdmin:
          type: boolean
        State:
          type: string
          enum: [active, passive, disconnected, killed]
        LastSeen:
          type: string
          format: date-time
        TaskCount:
          type: integer
...
    BroadcastRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
paths:
  /api/login:
    post:
      summary: Authenticate and obtain a JWT pair
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Token pair issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
                  user:
                    type: string
                  role:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/refresh:
    post:
      summary: Exchange a refresh token for a new JWT pair (one-time use)
      description: >-
        Refresh tokens are ROTATED: the presented token is consumed and the
        response carries a new access token AND a new refresh token. Clients
        must persist the replacement; replaying a consumed refresh token
        returns 401. Legacy refresh tokens without a jti rotate once and the
        replacement is rotatable.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New token pair (token + refresh_token; store both)
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refresh_token:
                    type: string
                    description: Replacement refresh token (the presented one is now invalid)
                  expires_in:
                    type: integer
        '401':
          description: Invalid, expired or already-consumed refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/status:
    get:
      summary: Operational telemetry (requires sessions:list)
      tags: [System]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Active sessions, listeners and uptime
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient permissions
...
  /api/sessions:
    get:
      summary: List active sessions
      tags: [Sessions]
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          required: false
          description: >-
            Optional time window in days (1-90): only sessions last seen
            within the window are returned. Omit for the full listing.
          schema:
            type: integer
            minimum: 1
            maximum: 90
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
...
  /api/cmd:
    post:
      summary: Execute command on agent
      tags: [Commands]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Command result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Command execution failed
  /api/broadcast:
    post:
      summary: Execute command on all agents
      tags: [Commands]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BroadcastRequest'
      responses:
        '200':
          description: Results from all agents
  /api/vault:
    get:
      summary: List or search credentials
      tags: [Vault]
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 256
          description: Search query (max 256 characters; longer queries answer 400)
      responses:
        '200':
          description: List of credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credential'
    post:
      summary: Store a credential
      tags: [Vault]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credential'
      responses:
        '200':
          description: Credential stored
    delete:
      summary: Delete a credential (vault:delete required, admin only)
      tags: [Vault]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: Credential ID (cred-...)
      responses:
        '200':
          description: Credential deleted
        '403':
          description: Insufficient permissions (vault:delete)
        '404':
          description: Credential not found
  /api/files:
    get:
      summary: List exfiltrated files
      tags: [Files]
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          required: false
          description: >-
            Optional time window in days (1-90): only loot records created
            within the window are returned. Omit for the full listing.
          schema:
            type: integer
            minimum: 1
            maximum: 90
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileRecord'
    post:
      summary: Upload a file
      tags: [Files]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                filename:
                  type: string
                module:
                  type: string
                data:
                  type: string
                  format: byte
      responses:
        '200':
          description: File stored
    delete:
      summary: Purge all captured files (files:delete required)
      description: >-
        Bulk wipe: removes every exfiltrated artifact everywhere it lives —
        blobs on disk, the current-run listing and the persisted file_records
        rows of previous runs. Requires the files:delete permission (the same
        capability the single-file route enforces, so no role gains a bulk
        action its single-file route lacked). Returns the number of records
        purged; the first hard error aborts the sweep.
      tags: [Files]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All loot purged
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: purged
                  purged:
                    type: integer
                    description: Number of records removed
        '401':
          description: Missing or invalid token
        '403':
          description: Insufficient permissions (files:delete required)
        '500':
          description: A blob refused to be removed; the sweep aborted
...
  /api/modules:
    get:
      summary: List available modules
      tags: [Modules]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Module'
    post:
      summary: Register a new module
      tags: [Modules]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Module'
      responses:
        '200':
          description: Module registered
  /api/modules/push:
    post:
      summary: Push a module to an agent
      tags: [Modules]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [module]
              properties:
                module:
                  type: string
                agent_id:
                  type: string
      responses:
        '200':
          description: Module pushed
  /api/socks:
    get:
      summary: List active SOCKS proxies
      tags: [Tunneling]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of proxies
    post:
      summary: Start a SOCKS proxy
      tags: [Tunneling]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                port:
                  type: integer
      responses:
        '200':
          description: Proxy started
    delete:
      summary: Stop a SOCKS proxy
      tags: [Tunneling]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Proxy stopped
  /api/portfwd:
    get:
      summary: List port forwards
      tags: [Tunneling]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of forwards
    post:
      summary: Start port forward
      tags: [Tunneling]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                local_port:
                  type: integer
                remote_host:
                  type: string
                remote_port:
                  type: integer
      responses:
        '200':
          description: Forward started
    delete:
      summary: Stop port forward
      tags: [Tunneling]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Forward stopped
...
  /api/operators:
    get:
      summary: List operators (admin only)
      tags: [Operators]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of operators
        '403':
          description: Admin role required
    post:
      summary: Create an operator (admin only)
      tags: [Operators]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username:
                  type: string
                password:
                  type: string
                  description: Plain password; stored as bcrypt hash
                role:
                  type: string
                  enum: [admin, operator, viewer, auditor]
      responses:
        '200':
          description: Operator created
...
  /api/audit:
    get:
      summary: Read the append-only audit trail (audit:read — admin and auditor)
      description: >-
        Every API call, auth event, task and lifecycle action is written to
        the audit_log table by middleware — this is its read API. Entries
        carry usernames and IPs in their detail field, hence the gate: the
        audit:read permission belongs to the admin and auditor roles (the
        permission existed in rbac.go since the early rounds — round 17
        finally wires an endpoint to it). The trail is append-only: there is
        no update or delete path. Newest entries first.
      tags: [System]
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          description: >-
            Page size (1-500). Default 500; the server clamps anything above
            the cap instead of trusting the client.
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: action
          in: query
          required: false
          description: >-
            Exact event-type filter (e.g. api_call, auth_failed, auth_denied,
            task). Max 64 characters; unknown values return an empty array,
            never an error.
          schema:
            type: string
            maxLength: 64
        - name: user
          in: query
          required: false
          description: >-
            Exact operator filter (round 18). Rows attributed to the given
            account only — system events (operator_id 0) are excluded under
            this filter, which is the honest semantics of "what did THIS
            account do". Attributability requires the round-18 middleware:
            authenticated calls write their real operator id; unknown
            usernames return an empty array, never an error.
          schema:
            type: string
            maxLength: 64
      responses:
        '200':
          description: Array of audit entries (empty array, not null, when nothing matches)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    action:
                      type: string
                    detail:
                      type: string
                    operator:
                      type: string
                      description: >-
                        Attributed account (round 18) — empty for system events
                        such as task/session lifecycle written with operator_id 0.
                    created:
                      type: string
                      format: date-time
        '400':
          description: Invalid limit or over-long action/user filter
        '401':
          description: Missing or invalid token
        '403':
          description: Admin role required
        '405':
          description: Only GET is supported
  /api/notes:
    get:
      summary: List operator notes for a session (collab:read — every role has it)
      tags: [Collaboration]
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of notes
        '400':
          description: Missing session_id
    post:
      summary: Add or update a note for a session
      tags: [Collaboration]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, content]
              properties:
                session_id:
                  type: string
                content:
                  type: string
      responses:
        '200':
          description: Note saved
  /api/lock:
    post:
      summary: Acquire an exclusive lock on a session
      tags: [Collaboration]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
      responses:
        '200':
          description: Lock acquired
  /api/profiles:
    get:
      summary: List saved engagement profiles
      tags: [Collaboration]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of profiles
    post:
      summary: Save an engagement profile
      tags: [Collaboration]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile saved
...